// payload/custom-auth/src/payload-types.ts
export interface User {
  id: string;
  externalId?: {
    authStrategies?:
      | {
          providerUserId?: string | null;
          accessToken?: string | null;
          refreshToken?: string | null;
          tokenType?: string | null;
          idToken?: string | null;
          tokenExpiry?: string | null;
          refreshTokenExpiry?: string | null;
          tokenIssuedAt?: string | null;
          authProvider?: 'google' | null;
          id?: string | null;
        }[]
      | null;
  };
  updatedAt: string;
  createdAt: string;
  email: string;
  username: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
  collection: 'users';
}